Move app.py to the correct location
